+++ library/core/src/main/java/com/google/android/exoplayer2/analytics/AnalyticsCollector.java
@@ -595,6 +595,7 @@ public class AnalyticsCollector
         listener -> listener.onTracksChanged(eventTime, trackGroups, trackSelections));
   }
 
+  @Deprecated
   @Override
   public final void onStaticMetadataChanged(List<Metadata> metadataList) {
     EventTime eventTime = generateCurrentPlayerMediaPeriodEventTime();