+++ testutils/src/main/java/com/google/android/exoplayer2/testutil/StubExoPlayer.java
@@ -375,6 +375,7 @@ public class StubExoPlayer extends BasePlayer implements ExoPlayer {
     throw new UnsupportedOperationException();
   }
 
+  @Deprecated
   @Override
   public List<Metadata> getCurrentStaticMetadata() {
     throw new UnsupportedOperationException();